refactor: extract the new template-parser package
