+++ client/shared/src/api/extension/extensionHostState.ts
@@ -2,6 +2,8 @@ import * as comlink from 'comlink'
 import { BehaviorSubject, Observable, ReplaySubject, Subject } from 'rxjs'
 import * as sourcegraph from 'sourcegraph'
 
+import { Context } from '@sourcegraph/template-parser'
+
 import { ConfiguredExtension } from '../../extensions/extension'
 import { SettingsCascade } from '../../settings/settings'
 import { ReferenceCounter } from '../../util/ReferenceCounter'
@@ -11,7 +13,6 @@ import { ExtensionViewer, ViewerUpdate } from '../viewerTypes'
 
 import { ExecutableExtension, observeActiveExtensions } from './activation'
 import { ExtensionCodeEditor } from './api/codeEditor'
-import { Context } from './api/context/context'
 import { ExtensionDocument } from './api/textDocument'
 import { ExtensionWorkspaceRoot } from './api/workspaceRoot'
 import { InitData } from './extensionHost'
