+++ client/shared/src/api/protocol/contribution.ts
@@ -1,7 +1,8 @@
 import { Primitive } from 'utility-types'
 
+import { Expression, TemplateExpression } from '@sourcegraph/template-parser'
+
 import { KeyPath } from '../client/services/settings'
-import { Expression, TemplateExpression } from '../extension/api/context/expr/evaluator'
 
 // NOTE: You must manually keep this file in sync with extension.schema.json#/properties/contributes (and possibly
 // extension_schema.go, if your changes are relevant to the subset of this schema used by our Go code).
