+++ client/template-parser/babel.config.js
@@ -1,0 +1,5 @@
+// @ts-check
+
+module.exports = {
+  extends: '../../babel.config.js',
+}
