+++ client/template-parser/jest.config.js
@@ -1,0 +1,13 @@
+// @ts-check
+
+const config = require('../../jest.config.base')
+
+const exportedConfig = {
+  ...config,
+  displayName: 'template-parser',
+  rootDir: __dirname,
+  roots: ['<rootDir>'],
+  verbose: true,
+}
+
+module.exports = exportedConfig
