+++ client/template-parser/package.json
@@ -1,0 +1,13 @@
+{
+  "private": true,
+  "name": "@sourcegraph/template-parser",
+  "version": "0.0.1",
+  "description": "Sourcegraph Template Parser",
+  "main": "./src/index.ts",
+  "sideEffects": false,
+  "license": "Apache-2.0",
+  "scripts": {
+    "eslint": "eslint --cache 'src/**/*.[jt]s?(x)'",
+    "test": "jest"
+  }
+}
