+++ client/template-parser/src/types.ts
@@ -1,0 +1,12 @@
+/**
+ * Context is an arbitrary, immutable set of key-value pairs. Its value can be any JSON object.
+ *
+ * @template T If you have a value with a property of type T that is not one of the primitive types listed below
+ * (or Context), you can use Context<T> to hold that value. T must be a value that can be represented by a JSON
+ * object.
+ */
+export interface Context<T = never>
+    extends Record<
+        string,
+        string | number | boolean | null | Context<T> | T | (string | number | boolean | null | Context<T> | T)[]
+    > {}
