+++ dev/foreach-ts-project.sh
@@ -19,6 +19,7 @@ DIRS=(
   client/common
   client/codeintellify
   client/wildcard
+  client/template-parser
   client/extension-api
   client/eslint-plugin-sourcegraph
   client/extension-api-types
