+++ tsconfig.all.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "references": [
+    { "path": "client/template-parser" },
     { "path": "client/build-config" },
     { "path": "client/common" },
     { "path": "client/shared" },
